feat: action.rs to action/
