+++ identity-wallet/src/state/actions/cancel_user_flow.rs
@@ -1,0 +1,20 @@
+use crate::reducer;
+use crate::state::actions::ActionTrait;
+use crate::state::actions::Reducer;
+use crate::state::reducers::cancel_user_flow;
+
+use ts_rs::TS;
+
+#[derive(serde::Serialize, serde::Deserialize, Debug, TS, Clone)]
+#[ts(export, export_to = "bindings/actions/CancelUserFlow.ts")]
+pub struct CancelUserFlow {
+    #[ts(optional)]
+    pub redirect: Option<String>,
+}
+
+#[typetag::serde(name = "[User Flow] Cancel")]
+impl ActionTrait for CancelUserFlow {
+    fn reducers<'a>(&self) -> Vec<Reducer<'a>> {
+        vec![reducer!(cancel_user_flow)]
+    }
+}
