+++ identity-wallet/src/state/actions/connection_accepted.rs
@@ -1,0 +1,17 @@
+use crate::reducer;
+use crate::state::actions::ActionTrait;
+use crate::state::actions::Reducer;
+use crate::state::reducers::authorization::handle_siopv2_authorization_request;
+
+use ts_rs::TS;
+
+#[derive(serde::Serialize, serde::Deserialize, Debug, TS, Clone)]
+#[ts(export, export_to = "bindings/actions/ConnectionAccepted.ts")]
+pub struct ConnectionAccepted;
+
+#[typetag::serde(name = "[Authenticate] Connection accepted")]
+impl ActionTrait for ConnectionAccepted {
+    fn reducers<'a>(&self) -> Vec<Reducer<'a>> {
+        vec![reducer!(handle_siopv2_authorization_request)]
+    }
+}
