+++ identity-wallet/src/state/actions/qr_code_scanned.rs
@@ -1,0 +1,19 @@
+use crate::reducer;
+use crate::state::actions::ActionTrait;
+use crate::state::actions::Reducer;
+use crate::state::reducers::{authorization::read_authorization_request, credential_offer::read_credential_offer};
+
+use ts_rs::TS;
+
+#[derive(serde::Serialize, serde::Deserialize, Debug, TS, Clone)]
+#[ts(export, export_to = "bindings/actions/QrCodeScanned.ts")]
+pub struct QrCodeScanned {
+    pub form_urlencoded: String,
+}
+
+#[typetag::serde(name = "[QR Code] Scanned")]
+impl ActionTrait for QrCodeScanned {
+    fn reducers<'a>(&self) -> Vec<Reducer<'a>> {
+        vec![reducer!(read_authorization_request), reducer!(read_credential_offer)]
+    }
+}
