+++ identity-wallet/src/state/actions/update_credential_metadata.rs
@@ -1,0 +1,28 @@
+use crate::reducer;
+use crate::state::actions::ActionTrait;
+use crate::state::actions::Reducer;
+use crate::state::reducers::update_credential_metadata;
+
+use ts_rs::TS;
+
+#[derive(serde::Serialize, serde::Deserialize, Debug, TS, Clone)]
+#[ts(export, export_to = "bindings/actions/UpdateCredentialMetadata.ts")]
+pub struct UpdateCredentialMetadata {
+    #[ts(type = "string")]
+    pub id: uuid::Uuid,
+    #[ts(optional)]
+    pub name: Option<String>,
+    #[ts(optional)]
+    pub icon: Option<String>,
+    #[ts(optional)]
+    pub color: Option<String>,
+    #[ts(optional)]
+    pub is_favorite: Option<bool>,
+}
+
+#[typetag::serde(name = "[Credential Metadata] Update")]
+impl ActionTrait for UpdateCredentialMetadata {
+    fn reducers<'a>(&self) -> Vec<Reducer<'a>> {
+        vec![reducer!(update_credential_metadata)]
+    }
+}
