+++ identity-wallet/src/state/actions/user_data_query.rs
@@ -1,0 +1,43 @@
+use crate::reducer;
+use crate::state::actions::ActionTrait;
+use crate::state::actions::Reducer;
+use crate::state::reducers::user_data_query::{connection_query, credential_query};
+
+use serde::{Deserialize, Serialize};
+use ts_rs::TS;
+
+#[derive(Clone, Serialize, Debug, Deserialize, TS, PartialEq)]
+#[ts(export)]
+pub enum QueryTarget {
+    Credentials,
+    Connections,
+}
+
+#[derive(Clone, Serialize, Debug, Deserialize, TS, PartialEq)]
+#[ts(export)]
+pub enum SortMethod {
+    NameAZ,
+    IssuanceNewOld,
+    AddedNewOld,
+    FirstInteractedNewOld,
+    LastInteractedNewOld,
+}
+
+#[derive(Serialize, Deserialize, Debug, TS, Clone)]
+#[ts(export, export_to = "bindings/actions/UserDataQuery.ts")]
+pub struct UserDataQuery {
+    pub target: QueryTarget,
+    #[ts(optional)]
+    pub search_term: Option<String>,
+    #[ts(optional)]
+    pub sort_method: Option<SortMethod>,
+    #[serde(default)]
+    pub sort_reverse: bool,
+}
+
+#[typetag::serde(name = "[User Data] Query")]
+impl ActionTrait for UserDataQuery {
+    fn reducers<'a>(&self) -> Vec<Reducer<'a>> {
+        vec![reducer!(credential_query), reducer!(connection_query)]
+    }
+}
