--- unime/src-tauri/.cargo/config.toml
@@ -1,4 +1,0 @@
-[build]
-target = "x86_64-unknown-linux-gnu"
-
-[target]
