+++ unime/src-tauri/Cargo.toml
@@ -8,13 +8,14 @@ keywords = ["identity", "did", "ssi", "wallet", "siopv2"]
 license = "Apache-2.0"
 repository = "https://github.com/impierce/identity-wallet"
 edition = "2021"
+rust-version = "1.75.0"
 
 [lib]
 name = "unime"
 crate-type = ["staticlib", "cdylib", "rlib"]
 
 [build-dependencies]
-tauri-build = { version = "=2.0.0-alpha.13", features = [] }
+tauri-build = { version = "=2.0.0-alpha.14", features = [] }
 
 [dependencies]
 tauri.workspace = true
@@ -39,7 +40,7 @@ wiremock.workspace = true
 serial_test.workspace = true
 tempfile.workspace = true
 
-tauri = { version = "=2.0.0-alpha.20", features = ["rustls-tls", "test"] }
+tauri = { version = "=2.0.0-alpha.21", features = ["rustls-tls", "test"] }
 serde = { version = "1.0", features = ["derive"] }
 tokio = { version = "1.26.0", features = ["macros"] }
 url = "2.4"
